Itemlock: user ids as long. Refs #25

+++ ComponentRegistry/src/test/java/clarin/cmdi/componentregistry/persistence/jpa/ItemLockDaoTest.java
@@ -49,7 +49,7 @@ public class ItemLockDaoTest extends BaseUnitTest {
 
     @Test
     public void testInsertLock() {
-        int userId = createUser().intValue();
+        long userId = createUser().longValue();
         int itemId = createItem().intValue();
 
         final ItemLock lock = new ItemLock();
@@ -68,7 +68,7 @@ public class ItemLockDaoTest extends BaseUnitTest {
 
     @Test(expected = DataIntegrityViolationException.class)
     public void testInsertLockTwice() {
-        int userId = createUser().intValue();
+        long userId = createUser().longValue();
         int itemId = createItem().intValue();
 
         final ItemLock lock = new ItemLock();
